requirements.txt
Django==4.1.5
djangorestframework==3.14.0
requests==2.28.2
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
django-cors-headers==3.13.0
django-rest-framework==0.1.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.2
openpyxl==3.1.0
pandas==1.5.3
pycountry==22.3.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0

djangorestframework-jwt~=1.11.0
PyJWT~=1.7.1
python-dotenv~=1.0.0
celery~=5.3.1
django-debug-toolbar~=4.2.0
django-filter~=23.2